requirements.txt
appdirs==1.4.4
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5.3
certifi==2019.11.28
certificate-gen==0.2
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
docutils==0.16
filelock==3.0.12
idna==2.9
isort==4.3.21
joblib==0.14.1
keyring==21.2.1
mccabe==0.6.1
packaging==20.4
pbr==5.4.5
Pillow==7.2.0
pkginfo==1.5.0.1
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2019.3
pywin32-ctypes==0.2.0
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.14.0
sqlparse==0.3.0
stevedore==3.2.0
toml==0.10.1
tqdm==4.48.0
twine==3.2.0
urllib3==1.25.10
webencodings==0.5.1
wrapt==1.11.2

